This is a large commit for files edited based on the 2016 DOE final report.
Main Calcs For Paper Script Edits:

--Libraries--
Added note that 64 bit Java is needed for the xlsx package.

Added the rootSolve library.
Deleted the GISTools library because it is not used.

Added a modification to the vioplot function, which allows for coloring each violin a different color

--Working Directories--
All working directories use the getwd() command now.

--File Loading--
The States and Counties file loads were changed to my computer's directories. This should be changed back to the original for general use.

Updated the epsg to 26917 for NAD83 UTM 17N.

rm(cities) added.

For buffering, used more quadsegs in the gBuffer command. The original made blocky buffers.

Added 3 color maps and files to the code.

--Risk Factors--
Thermal:
Added new thermal resource thresholds. Old ones are still in the file.
Updated the range of the means and the standard deviations for the interpolation tables.

Reservoir:
The reservoir risk factor now has 3 different objectives: RPIw, RPIg, and RFC.
The thresholds are different for these than before.
Some lines have been commented out because the most recent dataset did not have zeros in the reservoir data (no problems with taking the ln).
Updated the files to rm() for the reservoir analysis.

Utilization:
Wrote out FALSE and TRUE for F and T.
rm(utTemp) file.
Added a note that -Inf values are generated for NAs when using the focal() function

Seismic:
Using NormAngs now instead of the RiskAngs. These data are now read in separately.
Commented out lines related specifically to the RiskAng properties
Changed the range of the mean to [0, 180]

Combined:
There are many comb_extracts as a result of 3/5 color, 3 reservir objectives, and 2 thermal thresholds (new and old).

--Maps--
All maps now use the 'min' color option of makeMap() because all of the new maps are on [0,3] or [0, 5]

--Finding the point with the maximum value near a point of interest (POI):
Added that the search for maximum value is a 10 km radius.
Added 2 more objectives for selecting the best point:
1) Select all points that have the max value
2) Choose those points with the lowest uncertainty
3) Choose the point closest to the POI

--Monte Carlo Loops--
Deleted the counter variable, j, for colors. Used IndNA variable instead.
Made all thresholds variables, and indexed the values.
Normal angles are used for seismic. Also for seismic, new method to convert all angles to [0,180], and the error terms for the risk angle are computed.
Commented out some of the minimum distributions. These can be added back.
Geology only results are at the end

--Weibul Parameters--
Added comments for what params is

--Plots--
Boxplot, violin plot, and parallel axis plot are all made after each Monte Carlo loop.

--Comment-Out--
Commented out large segments of code above Scatter Plots of Metrics

makeMap.R Function Updates:
Updated the colorbar to make sure that all values are represented.
There was a bug that assigned white to every value not in the range of the colorbar.

Make Interp Tables Script Updates:
Updated the seismic angle uncertainty to reflect the use of Normal Angles.
All thresholds are referenced as variables and indexed.
Log space and real space are calculated in the same loop to save time.
3 and 5 color also in the same loop.