Modified the Weibull distribution fitting function to include a lower bound. The new file is called rw_functions_lb.R. The lb is the lower bound parameter, which is set in the main script as a matrix for all of the sites (columns are the favorability factors, rows are the sites).
Some sites have a sample mean that is close to the lower bound of the Weibull distribution. This throws a warning, and results in no fit for the site. For these sites, the shape parameter is fixed to 1.5, and the scale parameter is solved using the method of moments. The fit is not great, but it is better than before.

All of the main script has not been updated with the new set of functions yet. Only the first instance of the Weibull distribution fitting around line 5400 has been edited.