Updated the roots of the Weibull distribution to be 0 when the utilization risk factor is 0 with no uncertainty.
Uncommented many of the lines at the bottom of the code. These lines make figures that compare the metrics to each other. For the minimum metric, added the median of the Weibull distribution to one of these figures as plus signs.

Using the Weibull lower bound fitting with more of the combined metrics now.