Added the doubly truncated normal distribution and the truncated folded normal distribution to the rw_functions_lb.R script and the main_calcs_for_paper_CorrOverlay_CorrMeans.R script.
Corrected an error in computing the log-space variance of the Seismic variable from the real-space Taylor series.

Added a distribution fitting option in rw_functions_lb.R for Weibull to all variables, except when Seismic has a lower bound of 2.5. A doubly truncated normal is used for Seismic in that case. This distribution option was used for the RFC Average and RFC Geometric Mean.

Added computations of the Pearson correlation coefficients for the scatter plots.

Names of the places (e.g. Corning, NY) are now assigned to a vector.